+++ backend/canisters/local_user_index/api/src/updates/c2c_delete_group.rs
@@ -1,0 +1,16 @@
+use candid::CandidType;
+use oc_error_codes::OCError;
+use serde::{Deserialize, Serialize};
+use types::ChatId;
+
+#[derive(CandidType, Serialize, Deserialize, Debug)]
+pub struct Args {
+    pub chat_id: ChatId,
+}
+
+#[derive(CandidType, Serialize, Deserialize, Debug)]
+pub enum Response {
+    Success,
+    ChatNotFound,
+    Error(OCError),
+}