+++ backend/canisters/local_user_index/api/src/updates/c2c_local_group_index.rs
@@ -1,0 +1,10 @@
+use crate::LocalGroupIndexEvent;
+use serde::{Deserialize, Serialize};
+use types::{IdempotentEnvelope, UnitResult};
+
+#[derive(Serialize, Deserialize, Debug)]
+pub struct Args {
+    pub events: Vec<IdempotentEnvelope<LocalGroupIndexEvent>>,
+}
+
+pub type Response = UnitResult;